refactor(widget-chat): Restructure files, folders to match rest of sdk
