+++ packages/widget-chat/.storybook/config.js
@@ -1,7 +1,7 @@
-import { configure } from '@kadira/storybook';
+import {configure} from '@kadira/storybook';
 
 function loadStories() {
-  require('../stories');
+  require(`../stories`);
 }
 
 configure(loadStories, module);
