+++ packages/widget-chat/src/app.js
