+++ packages/widget-chat/src/components/activity-title/index.js
