+++ packages/widget-chat/src/containers/chat-widget/index.test.js
@@ -1,7 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import ChatWidget from './chat-widget';
-import TestUtils from 'react-addons-test-utils';
+import ChatWidget from '.';
+import {findRenderedDOMComponentWithTag, renderIntoDocument} from 'react-addons-test-utils';
 
 it(`is rendered properly`, () => {
   const widget = TestUtils.renderIntoDocument(
