+++ packages/widget-chat/src/containers/chat-widget/styles.css
