+++ packages/widget-chat/src/index.html
@@ -1,8 +1,8 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Chat Widget</title>
-    <meta charset="utf8"/>    
+    <title>Chat Widget Demo</title>
+    <meta charset="utf8"/>
   </head>
   <body>
     <main id="main"></main>
