+++ packages/widget-chat/webpack.config.js
@@ -32,7 +32,7 @@ process.env.NODE_ENV = process.env.NODE_ENV || 'development';
 module.exports = {
   context: __dirname,
   entry: [
-    './src/main.js'
+    './src/app.js'
   ],
   output: {
     filename: 'bundle.js',
